+++ asdf/asdf.py
@@ -9,6 +9,7 @@ import copy
 import datetime
 import warnings
 from pkg_resources import parse_version
+from urllib.parse import urlparse
 
 import numpy as np
 from jsonschema import ValidationError
@@ -1487,11 +1488,22 @@ def open_asdf(fd, uri=None, mode=None, validate_checksums=False,
         validate_on_read=validate_on_read)
 
 
+# astropy.io.fits supports opening files that are externally
+# compressed with gzip or zip, and asdf does not, so we may as
+# well give those extensions to FITS.
+_FITS_EXTENSIONS = [".fits", ".gz", ".zip"]
+_ASDF_EXTENSIONS = [".asdf"]
+
+
 def is_asdf_file(fd):
     """
     Determine if fd is an ASDF file.
 
-    Reads the first five bytes and looks for the ``#ASDF`` string.
+    For most input, reads the first five bytes and looks
+    for the ``#ASDF`` string.
+
+    For URL input, looks for an extension that should be passed
+    to AsdfInFits, otherwise assumes ASDF.
 
     Parameters
     ----------
@@ -1501,6 +1513,26 @@ def is_asdf_file(fd):
     if isinstance(fd, generic_io.InputStream):
         # If it's an InputStream let ASDF deal with it.
         return True
+
+    if isinstance(fd, str):
+        parsed = urlparse(fd)
+        if parsed.scheme in ["http", "https"]:
+            # We don't want to read URL content here because
+            # that will cause the file to be downloaded twice.
+            ext = os.path.splitext(parsed.path)[1].lower()
+            if ext in _FITS_EXTENSIONS:
+                return False
+            elif ext in _ASDF_EXTENSIONS:
+                return True
+            else:
+                message = (
+                    "The URL '{}' does not include an obvious FITS "
+                    "or ASDF filename extension.  Assuming ASDF.\n\n"
+                    "If this URL returns FITS content, it cannot be opened "
+                    "with asdf.open().  Use asdf.fits_embed.AsdfInFits.open() instead."
+                ).format(fd)
+                warnings.warn(message, AsdfWarning)
+                return True
 
     to_close = False
     if isinstance(fd, AsdfFile):